RayRewardServer/requirements.txt
ray[default]>=2.9.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyyaml>=6.0.0
numpy>=1.24.0
torch>=2.1.0
tensordict>=0.3.0
transformers>=4.40.0
vllm>=0.4.0
codetiming
omegaconf>=2.3.0
